Applied IIFE and @namespace, @memberOf
Getting jsdoc working initially with IIFE was pretty difficult, as it wasn't clear how to make make jsdoc compile the docs. Setting @namespace and @memberOf for each function helped clear that problem up, and I think it should be included in this example to save time for other people using this style guide and trying to figure out how to configure jsdoc here. 

If you do it a different way, do let me know how!